Now tests itself.
